requirements.txt
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
distro==1.9.0
docstring_parser==0.18.0
durationpy==0.10
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
jiter==0.14.0
kubernetes==35.0.0
markdown-it-py==4.2.0
mdurl==0.1.2
numpy==2.4.4
oauthlib==3.3.1
ollama==0.6.2
portalocker==3.2.0
protobuf==7.34.1
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qdrant-client==1.17.1
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
websocket-client==1.9.0
